изменил минимум маштаб на 25
