+++ js/main.js
@@ -2,4 +2,3 @@ import { similarPost } from './util.js';
 import { renderPicture } from './render-picture.js';
 renderPicture(similarPost);
 import './work-form.js';
-import './scale.js';
